answer Daniel's comments
